Fixed a bunch of bugs.
The script would previously attempt to send a message to everyone on the list sequentially, meaning that a failure at some point would leave some people with emails while the rest
would not get them. Now the script constructs the whole list first and only then attempts to send the emails.

Also fixed non-ascii characters handling in the regex.